+++ PlayerHandler.py
@@ -2,9 +2,7 @@ from player import MainWindow
 
 
 class PlayerHandler(MainWindow):
-    def open_error_corrector(self, checked):
-        if self.ErrorCorrector.isVisible():
-            self.ErrorCorrector.hide()
-        else:
-            self.ErrorCorrector.show()
+    def __init__(self):
+        super().__init__()
+
